Refactor "Blocks" to own module
